modify addExperttosubService and createSubService methods in subserviceController

+++ src/main/java/ir/maktab/homeservicesystem/dto/mapper/SubServiceCreateParam.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class SubServiceCreateParam {
     private String name;
-    private long mainCategoryId;
+    private Integer mainCategoryId;
 }
